frontend: Rename Tab to Tabs
