+++ frontend/src/components/common/index.ts
@@ -14,8 +14,8 @@ export * from './SectionHeader';
 export { default as SectionHeader } from './SectionHeader';
 export * from './SimpleTable';
 export { default as SimpleTable } from './SimpleTable';
-export * from './Tab';
-export { default as Tab } from './Tab';
+export * from './Tabs';
+export { default as Tabs } from './Tabs';
 export * from './Terminal';
 export { default as Terminal } from './Terminal';
 export * from './Tooltip';
